refactor(actions/tasks): co-locate task actions in subfolders
