+++ src/actions/tasks/delete-task/delete-task.ts
@@ -2,7 +2,7 @@
 
 import { auth } from "@/auth"
 import { db } from "@/lib/db"
-import { revalidateTaskPaths, type ActionResult } from "./shared"
+import { revalidateTaskPaths, type ActionResult } from "../shared"
 
 export async function deleteTask(
   taskId: string
