+++ src/actions/tasks/index.ts
@@ -1,6 +1,5 @@
-export { addTaskTag } from "./add-task-tag"
-export { createTask } from "./create-task"
-export { deleteTask } from "./delete-task"
-export { duplicateTask } from "./duplicate-task"
-export { updateTask } from "./update-task"
-export { updateTaskStatus } from "./update-task-status"
+export { addTaskTag } from "./add-task-tag/add-task-tag"
+export { createTask } from "./create-task/create-task"
+export { deleteTask } from "./delete-task/delete-task"
+export { duplicateTask } from "./duplicate-task/duplicate-task"
+export { updateTask, updateTaskStatus } from "./update-task/update-task"
