+++ src/actions/tasks/update-task/update-task.ts
@@ -2,13 +2,13 @@
 
 import { auth } from "@/auth"
 import { db } from "@/lib/db"
-import { updateTaskSchema } from "@/validations/task"
+import { updateTaskSchema, updateTaskStatusSchema } from "@/validations/task"
 import {
   parseDatetime,
   parseTags,
   revalidateTaskPaths,
   type ActionResult,
-} from "./shared"
+} from "../shared"
 
 export async function updateTask(
   taskId: string,
@@ -65,6 +65,40 @@ export async function updateTask(
           ? parseDatetime(parsed.data.dueDate)
           : undefined,
     },
+  })
+
+  revalidateTaskPaths()
+  return { success: true, data: undefined }
+}
+
+export async function updateTaskStatus(
+  taskId: string,
+  formData: FormData
+): Promise<ActionResult<void>> {
+  const session = await auth()
+  if (!session?.user?.id) {
+    return { success: false, error: "Unauthorized" }
+  }
+
+  const parsed = updateTaskStatusSchema.safeParse({
+    status: formData.get("status"),
+  })
+  if (!parsed.success) {
+    return { success: false, error: parsed.error.errors[0].message }
+  }
+
+  const existing = await db.task.findUnique({
+    where: { id: taskId, userId: session.user.id },
+    select: { id: true },
+  })
+
+  if (!existing) {
+    return { success: false, error: "Task not found" }
+  }
+
+  await db.task.update({
+    where: { id: taskId, userId: session.user.id },
+    data: { status: parsed.data.status },
   })
 
   revalidateTaskPaths()
